resolve error the ts

+++ src/controllers/products/PATCH/changeImageProduct.ts
@@ -1,5 +1,5 @@
 import { Request, Response } from "express";
-import { uploadToCloudinary, db, deleteImageToCloudinary } from "@lib";
+import { uploadToCloudinary, models, deleteImageToCloudinary } from "@lib";
 
 const Product = models.Product;
 
@@ -16,7 +16,12 @@ export const changeImageProduct = async (req: Request, res: Response) => {
       .status(400)
       .json({ error: true, message: "Error al subir la imagen" });
   const product = await Product.findByPk(req.params.id);
-  await deleteImageToCloudinary(`${userId}/${product.primary_image}`);
-  await product.update({ primary_image: image_url });
-  return res.status(200).json({ error: false });
+  if (product) {
+    await deleteImageToCloudinary(`${userId}/${product.primary_image}`);
+    await product.update({ primary_image: image_url });
+    return res.status(200).json({ error: false });
+  }
+  return res
+    .status(400)
+    .json({ error: true, message: "Error al subir la imagen" });
 };
